# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AdminCreateUserRequest:
    properties:
      avatar:
        type: string
      email:
        type: string
      is_active:
        description: 使用指针以区分 "未提供" 和 "设置为false"
        type: boolean
      password:
        minLength: 6
        type: string
      phone:
        type: string
      real_name:
        type: string
      role_ids:
        description: 关联的角色ID列表
        items:
          type: string
        type: array
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  dto.AdminUpdateUserRequest:
    properties:
      avatar:
        type: string
      email:
        type: string
      is_active:
        type: boolean
      phone:
        type: string
      real_name:
        type: string
      role_ids:
        items:
          type: string
        type: array
    type: object
...
  dto.RoleCreateRequest:
    properties:
      description:
        type: string
      display_name:
        maxLength: 100
        minLength: 2
        type: string
      name:
        maxLength: 50
        minLength: 2
        type: string
    required:
    - display_name
    - name
    type: object
  dto.RoleResponse:
    properties:
      description:
        type: string
      display_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    type: object
  dto.RoleUpdateRequest:
    properties:
      description:
        type: string
      display_name:
        maxLength: 100
        minLength: 2
        type: string
      is_active:
        type: boolean
    required:
    - display_name
    type: object
  dto.UserListResponse:
    properties:
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
    type: object
  dto.UserResponse:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      is_active:
        type: boolean
      phone:
        type: string
      real_name:
        type: string
      roles:
        items:
          type: string
        type: array
      username:
        type: string
      uuid:
        type: string
    type: object
...
  resp.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a lightweight CRM system.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: CRM Lite API
  version: "1.0"
paths:
...
  /roles:
    get:
      consumes:
      - application/json
      description: 获取所有可用的用户角色
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.RoleResponse'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取角色列表
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: 创建一个新的用户角色
      parameters:
      - description: 角色信息
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/dto.RoleCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.RoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建角色
      tags:
      - Roles
  /roles/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除一个角色
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除角色
      tags:
      - Roles
    get:
      consumes:
      - application/json
      description: 根据角色ID获取详细信息
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.RoleResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取单个角色详情
      tags:
      - Roles
    put:
      consumes:
      - application/json
      description: 更新一个已存在角色的信息
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: 要更新的角色信息
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/dto.RoleUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.RoleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新角色
      tags:
      - Roles
...
  /users:
    get:
      consumes:
      - application/json
      description: 分页、筛选、搜索用户列表
      parameters:
      - description: 按邮箱搜索
        in: query
        name: email
        type: string
      - description: 按状态筛选
        in: query
        name: is_active
        type: boolean
      - in: query
        name: page
        type: integer
      - in: query
        name: page_size
        type: integer
      - description: 按用户名模糊搜索
        in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取用户列表
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: 由管理员创建一个新的用户账号并可以指定角色
      parameters:
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.AdminCreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: 管理员创建用户
      tags:
      - Users
  /users/{uuid}:
    delete:
      consumes:
      - application/json
      description: 管理员删除一个用户
      parameters:
      - description: User UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: 删除用户
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: 根据用户UUID获取详细信息
      parameters:
      - description: User UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取单个用户详情
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: 管理员更新用户信息，包括角色等
      parameters:
      - description: User UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: 要更新的用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.AdminUpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - ApiKeyAuth: []
      summary: 管理员更新用户
      tags:
      - Users
swagger: "2.0"